requirements.txt
django
psycopg2-binary
djangorestframework
djangorestframework_simplejwt
flake8
django-cors-headers
python-dateutil
